agents/incident/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv 